changed app logic realisation from double to BigDecimal

+++ src/ru/netology/bookkeeping/Bill.java
@@ -2,20 +2,22 @@ package ru.netology.bookkeeping;
 
 import ru.netology.bookkeeping.taxtypes.TaxType;
 
+import java.math.BigDecimal;
+
 public class Bill {
 
-    private double amount;
+    private BigDecimal amount;
     private TaxType taxType;
     private TaxService taxService;
 
     public Bill(double amount, TaxType taxType, TaxService taxService) {
-        this.amount = amount;
+        this.amount = BigDecimal.valueOf(amount);
         this.taxType = taxType;
         this.taxService = taxService;
     }
 
     public void payTaxes() {
-        double taxAmount = taxType.calculateTaxFor(amount);
+        BigDecimal taxAmount = taxType.calculateTaxFor(amount);
         taxService.payOut(taxAmount);
     }
 }
